% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFST.R
\name{calcFST}
\alias{calcFST}
\title{Calculates SNP-level FST values given two populations}
\usage{
calcFST(genotype_file, fam_file, output_folder, out_file)
}
\arguments{
\item{genotype_file}{(char) path to PLINK-formatted SNP genotype data.}

\item{fam_file}{(char) path to PLINK population-coded fam file.}

\item{output_folder}{(char) path to output directory.}

\item{out_file}{(char) path to final GSEA-formatted output file.}
}
\value{
none
}
\description{
Calculates SNP-level FST values given two populations
}
