% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBiogrid.R
\name{getBiogrid}
\alias{getBiogrid}
\title{Script to determine the fraction of within and between-pathway interaction
discoveries supported by literature-based evidence (i.e., BioGRID)}
\usage{
getBiogrid(bgridF, inDir)
}
\arguments{
\item{bgridF}{(char) path to BioGRID human interaction file.}

\item{geneF}{(char) path to selection-enriched gene file (written by writePathFiles.R).}

\item{outDir}{(char) path to write output files.}
}
\value{
none
}
\description{
Script to determine the fraction of within and between-pathway interaction
discoveries supported by literature-based evidence (i.e., BioGRID)
}
