% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDstatsBPM.R, R/LDstatsWPM.R
\name{LDstats}
\alias{LDstats}
\title{Calculate selection statistics (LD) and perform exploratory analyses
for two sets of variants via R snpStats package}
\usage{
LDstats(hcInDir, lcInDir, statistic = "R.squared", popNames = NULL,
  outDir)

LDstats(hcInDir, lcInDir, statistic = "R.squared", popNames = NULL,
  outDir)
}
\arguments{
\item{hcInDir}{(char) path to files with high-confidence pathway SNP lists}

\item{lcInDir}{(char) path to files with low-confidence pathway SNP lists}

\item{statistic}{(char) function will generate both R squared and D prime
LD statistics; choose either 'R.squared' or 'D.prime' for plotting and
p value calculation (default="R.squared")
NOTE: syntax is important!}

\item{popNames}{(char) optional - character vector to set alternate
population names in plots eg. c("European", "African") rather than
"CEU" and "ASW"}

\item{outDir}{(char) path to output directory}

\item{snp2geneF}{(char) path to file with snp2gene mappings. Output of
mapSNP2gene() (found in GWAS2Pathway)}

\item{hcInDir}{(char) path to files with high-confidence pathway SNP lists}

\item{lcInDir}{(char) path to files with low-confidence pathway SNP lists}

\item{outDir}{(char) path to output directory}
}
\value{
none

none
}
\description{
Calculate selection statistics (LD) and perform exploratory analyses
for two sets of variants via R snpStats package

Calculate selection statistics (LD) and perform exploratory analyses
for two sets of variants via R snpStats package
https://bioconductor.org/packages/release/bioc/manuals/snpStats/man/snpStats.pdf
}
