% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDstatsBPM.R
\name{LDstatsBPM}
\alias{LDstatsBPM}
\title{Calculate selection statistics (LD) and perform exploratory analyses
for two sets of variants via R snpStats package}
\usage{
LDstatsBPM(hcInDir, lcInDir, snp2geneF, outDir)
}
\arguments{
\item{hcInDir}{(char) path to files with high-confidence pathway SNP lists}

\item{lcInDir}{(char) path to files with low-confidence pathway SNP lists}

\item{snp2geneF}{(char) path to file with snp2gene mappings. Output of
mapSNP2gene() (found in GWAS2Pathway)}

\item{outDir}{(char) path to output directory}
}
\value{
none
}
\description{
Calculate selection statistics (LD) and perform exploratory analyses
for two sets of variants via R snpStats package
}
