% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDstatsWPM.R
\name{LDstatsWPM}
\alias{LDstatsWPM}
\title{Calculate selection statistics (LD) and perform exploratory analyses
for two sets of variants via R snpStats package}
\usage{
LDstatsWPM(enrichDir, unenrichDir, pop1, pop2, outDir)
}
\arguments{
\item{enrichDir}{(char) path to selection-enriched pathway SNP lists}

\item{unenrichDir}{(char) path to unenriched pathway SNP lists}

\item{pop1}{(char) character code for the first population (controls).}

\item{pop2}{(char) character code for the second population (cases).}

\item{outDir}{(char) path to output directory}
}
\value{
none
}
\description{
Calculate selection statistics (LD) and perform exploratory analyses
for two sets of variants via R snpStats package
}
