% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recodeFAM.R
\name{recodeFAM}
\alias{recodeFAM}
\title{Recodes PLINK fam file with population-specific codes
adapted from SP plink_baseSetup.R (part of GWAS2pathway package)}
\usage{
recodeFAM(
  genotype_file,
  pop_one,
  pop_two,
  population_table,
  SET_SEED = 42,
  out_file = "population-coded"
)
}
\arguments{
\item{genotype_file}{(char) path to file with SNP genotype data (PLINK format)}

\item{pop_one}{(char) character code for the first population.}

\item{pop_two}{(char) character code for the second population.}

\item{population_table}{(char) path to file with population information.
Gives the number of samples per population in the dataset.}

\item{SET_SEED}{(integer) value for set.seed() before shuffling (default=42).}

\item{out_file}{(char) optional; name for fam file (default=population-coded).
file extension added.}
}
\value{
none
}
\description{
Recodes PLINK fam file with population-specific codes
adapted from SP plink_baseSetup.R (part of GWAS2pathway package)
}
