% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNP2gene.R
\name{SNP2gene}
\alias{SNP2gene}
\title{Maps SNPs to their nearest genes}
\usage{
SNP2gene(in_file, refgene_file, marg = 0, out_file)
}
\arguments{
\item{in_file}{(char) path to PLINK bim file (only CHR, SNP, and BP
columns considered).}

\item{refgene_file}{(char) path to refseq table with header.}

\item{marg}{(integer) region upstream and downstream of [txStart,txEnd] (default=0).}

\item{out_file}{(char) path to write snp-to-gene mapping file.}
}
\value{
none
}
\description{
NOTE: updated GenomicRanges to version 1.30.1 on 01/30/2018
Requires ignore.strand=TRUE param to properly run distanceToNearest()
given unknown strand assignment for SNP array-based genotyping info
}
