% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeEmapF.R
\name{writeEmapF}
\alias{writeEmapF}
\title{Write enrichments file for use in Cytoscape EnrichmentMap}
\usage{
writeEmapF(resEmF, enrichNES = 0.3, outF)
}
\arguments{
\item{resEmF}{(list) file paths to results_forEM.txt for each population comparison
(generated by setupGSEA_run.R).}

\item{enrichNES}{(integer) NES cutoff to select validated selection-enriched
pathways (default=0.3)}

\item{outF}{(char) name of output file.}
}
\value{
none. Output file write into directory of first population comparison.
   E.g., CEU_YRI directory
}
\description{
Write enrichments file for use in Cytoscape EnrichmentMap
}
