% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeEmapFile.R
\name{writeEmapFile}
\alias{writeEmapFile}
\title{Write enrichments file for use in Cytoscape EnrichmentMap}
\usage{
writeEmapFile(results_file, ENRICH_NES, out_file)
}
\arguments{
\item{results_file}{(list) file paths to results.txt for each population comparison
(generated by setupGSEA_run.R).}

\item{ENRICH_NES}{(integer) NES cutoff to select validated selection-enriched
pathways}

\item{out_file}{(char) name of output file.}
}
\value{
none. Output file write into directory of first population comparison.
   E.g., CEU_YRI directory
}
\description{
Write enrichments file for use in Cytoscape EnrichmentMap
}
