% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFST.R
\name{calcFST}
\alias{calcFST}
\title{Calculates SNP-level FST values for use in setupGSEArun.R}
\usage{
calcFST(genotype_file, fam_file, output_folder, out_file)
}
\arguments{
\item{genotype_file}{(char) path to file with SNP genotype data (PLINK format).}

\item{fam_file}{(char) path to PLINK population coded fam file.}

\item{output_folder}{(char) path to directory to store output FST files generated by PLINK.}

\item{out_file}{(char) path to final SNP FST file used for GSEA input.}
}
\value{
none
}
\description{
Calculates SNP-level FST values for use in setupGSEArun.R
}
