% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFST.R
\name{calcFST}
\alias{calcFST}
\title{Calculates SNP-level FST values for use in setupGSEArun.R}
\usage{
calcFST(genoF, realFam, outDir, outF)
}
\arguments{
\item{genoF}{(char) path to file with SNP genotype data (PLINK format).}

\item{realFam}{(char) path to PLINK case/control coded fam file.}

\item{outDir}{(char) directory to store output files.}

\item{outF}{(char) path to write SNP-FST file.}
}
\value{
none
}
\description{
Calculates SNP-level FST values for use in setupGSEArun.R
}
