% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFST.R
\name{calcFST}
\alias{calcFST}
\title{Calculates SNP-level FST values for use in setupGSEArun.R}
\usage{
calcFST(genoF, famF, outDir)
}
\arguments{
\item{genoF}{(char) path to file with SNP genotype data (PLINK format).}

\item{famF}{(char) path to PLINK population coded fam file.}

\item{outDir}{(char) path to directory to store output FST files generated by PLINK.}

\item{outF}{(char) path to final SNP FST file used for GSEA input.}
}
\value{
none
}
\description{
Calculates SNP-level FST values for use in setupGSEArun.R
}
