% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEmap.R
\name{plotEmap}
\alias{plotEmap}
\title{Create EnrichmentMap in Cytoscape to visualize selection-enriched pathways}
\usage{
plotEmap(gmtF, eMapF, netName = "generic", imageFormat = "png")
}
\arguments{
\item{gmtF}{(char) file path to GMT file (this needs to be absolute path name).}

\item{eMapF}{(char) file path to results_forEM_selEnrich.txt
(generated by writeEmapFile.R).}

\item{netName}{(char) name for network in Cytoscape.}

\item{imageFormat}{(char) one of PNG, PDF, SVG, or JPEG.

# not run because requires Cytoscape to be installed and open
# plotEmap(gmtFile = gmtF, nodeAttrFile = nodeAttrFile, netName="HighRisk",
# outDir=outDir).}
}
\value{
Filename of image to which EnrichmentMap is exported. Also side
		effect of plotting the EnrichmentMap in an open session of Cytoscape.
}
\description{
Create a network where nodes are pathway gene sets significantly
	enriched in two independent population-based GSEA (i.e., validated
		selection-enriched gene sets.) Pathways are then clustered to identify
	main pathways and themes of the selection-enriched gene sets.
}
