% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popPCA.R
\name{popPCA}
\alias{popPCA}
\title{Calculate population stratification between different
ancestry groups (e.g., CEU vs YRI) via complete linkage clustering}
\usage{
popPCA(genoF, famF, pop1, pop2, dimensions = 3L, outF)
}
\arguments{
\item{genoF}{(char) path to file with SNP genotype data (PLINK format).}

\item{famF}{(char) path to PLINK fam file (case/control population coded).}

\item{pop1}{(char) character code for the first population (controls).}

\item{pop2}{(char) character code for the second population (cases).}

\item{dimensions}{(integer, default=3).}

\item{outF}{(char) path to write PNG image.}
}
\value{
none
}
\description{
Calculate population stratification between different
ancestry groups (e.g., CEU vs YRI) via complete linkage clustering
}
