% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupGSEArun.R
\name{setupGSEArun}
\alias{setupGSEArun}
\title{Sets up and runs GSEA using population-based FST
table containing GSEA results (Geneset, Size, ES, NES, NominalP, FDR, FWER)}
\usage{
setupGSEArun(realF, pathF, snp2geneF, setPerm = 1000L,
  snp2genedist = 500000L, minGene = 10L, maxGene = 300L,
  setSeed = 42L, outDir)
}
\arguments{
\item{realF}{(char) path to file with real SNP association statistics}

\item{pathF}{(char) path to pathway definitions GMT file}

\item{snp2geneF}{(char) path to file with snp2gene mappings. Output of
mapSNP2gene() (found in GWAS2Pathway).}

\item{setPerm}{(integer) set cycle of permutations to run
default=1000}

\item{snp2genedist}{(integer) value for GSEA --distance.
Max. distance between SNP and gene for their association in snp2geneF}

\item{minGene}{(integer) value for GSEA --setmin.
Min. number of genes in a gene set to be considered}

\item{maxGene}{(integer) value for GSEA --setmax.
Max. number of genes in a gene set to be considered}

\item{setSeed}{(integer) value for GSEA --seed}
}
\value{
none
}
\description{
Sets up and runs GSEA using population-based FST
table containing GSEA results (Geneset, Size, ES, NES, NominalP, FDR, FWER)
}
