% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNP2gene.R
\name{SNP2gene}
\alias{SNP2gene}
\title{Maps SNPs to their nearest genes (for use in setupGSEArun.R)
(adapted from map_SNP2gene from SP's GWAS2pathway package)}
\usage{
SNP2gene(inF, geneF, marg = 0L, outDir)
}
\arguments{
\item{inF}{(char) path to PLINK .bim file (only CHR, SNP, and BP
columns considered.}

\item{geneF}{(char) path to refseq table with header.}

\item{marg}{(integer) region upstream and downstream of [txStart,txEnd].}

\item{outDir}{(char) path to write snp-to-gene mapping.}
}
\value{
none
}
\description{
NOTE: updated GenomicRanges to version 1.30.1 on 01/30/2018
Requires ignore.strand=TRUE param to properly run distanceToNearest()
given unknown strand assignment for SNP array-based genotyping info
}
