% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNPassocBPM.R
\name{SNPassocBPM}
\alias{SNPassocBPM}
\title{Calculate selection statistics (LD) and perform exploratory analyses
for two sets of variants via R snpStats package}
\usage{
SNPassocBPM(enrichDir, unenrichDir, pop1, pop2, snp2geneF, outDir)
}
\arguments{
\item{enrichDir}{(char) path to selection-enriched pathway SNP lists}

\item{unenrichDir}{(char) path to unenriched pathway SNP lists}

\item{pop1}{(char) character code for the first population (controls).}

\item{pop2}{(char) character code for the second population (cases).}

\item{snp2geneF}{(char) path to file with snp2gene mappings. Output of
mapSNP2gene() (found in GWAS2Pathway)}

\item{outDir}{(char) path to output directory}
}
\value{
none
}
\description{
Calculate selection statistics (LD) and perform exploratory analyses
for two sets of variants via R snpStats package
}
