% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNPassocWPM.R
\name{SNPassocWPM}
\alias{SNPassocWPM}
\title{Calculates SNP-SNP association statistics and perform exploratory analyses
within pathways}
\usage{
SNPassocWPM(
  pop_one,
  pop_two,
  snp2gene_file,
  ASSOC_FDR = 0.05,
  enrich_folder,
  unenrich_folder,
  output_folder
)
}
\arguments{
\item{pop_one}{(char) character code for the first population (e.g., CEU).}

\item{pop_two}{(char) character code for the second population (e.g., YRI).}

\item{snp2gene_file}{(char) path to file with snp-to-gene mappings
(output of SNP2gene.R).}

\item{ASSOC_FDR}{(interger) FDR value to define significant pathway coevolution
(default=0.05).}

\item{enrich_folder}{(char) path to selection-enriched pathway SNP lists.}

\item{unenrich_folder}{(char) path to unenriched pathway SNP lists.}

\item{output_folder}{(char) path to output directory.}
}
\value{
none
}
\description{
Calculates SNP-SNP association statistics and perform exploratory analyses
within pathways
}
