% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNPassocWPM.R
\name{SNPassocWPM}
\alias{SNPassocWPM}
\title{Calculate selection statistics (LD) and perform exploratory analyses
for two sets of variants via R snpStats package}
\usage{
SNPassocWPM(
  pop_one,
  pop_two,
  ASSOC_FDR,
  enrich_folder,
  unenrich_folder,
  output_folder
)
}
\arguments{
\item{pop_one}{(char) character code for the first population.}

\item{pop_two}{(char) character code for the second population.}

\item{ASSOC_FDR}{(interger) FDR value to define significant pathway coevolution.}

\item{enrich_folder}{(char) path to selection-enriched pathway SNP lists}

\item{unenrich_folder}{(char) path to unenriched pathway SNP lists}

\item{output_folder}{(char) path to output directory}
}
\value{
none
}
\description{
Calculate selection statistics (LD) and perform exploratory analyses
for two sets of variants via R snpStats package
}
