% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeEmapFile.R
\name{writeEmapFile}
\alias{writeEmapFile}
\title{Write enrichments file for use in Cytoscape EnrichmentMap}
\usage{
writeEmapFile(resEmF, enrichNES = 0.3, outDir)
}
\arguments{
\item{resEmF}{(list) file paths to results_forEM.txt for each population comparison
(generated by setupGSEA_run.R).}

\item{enrichNES}{(integer) NES cutoff to select validated selection-enriched
pathways (default=0.3)}

\item{outDir}{(char) path to directory where file should be stored.}
}
\value{
none
}
\description{
Write enrichments file for use in Cytoscape EnrichmentMap
}
