% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writePathFiles.R
\name{writePathFiles}
\alias{writePathFiles}
\title{Generates SNP lists per selection-enriched and unenriched pathway
as determined by GSEA}
\usage{
writePathFiles(genoF, resF, gseaStatF, snp2geneF, famF, enrichNES = 0.3,
  unenrichNES = 0.1, enrichDir, unenrichDir)
}
\arguments{
\item{genoF}{(char) path to file with SNP genotype data (PLINK format).}

\item{resF}{(char) path to files with GSEA results.
Strutured to compare results of two population analyses
i.e., CEU vs. YRI and CEU vs. LWK.}

\item{gseaStatF}{(char) path to GSEA statistics file.}

\item{snp2geneF}{(char) path to SNP-gene mapping file.}

\item{famF}{(char) path to PLINK population coded fam file.}

\item{enrichNES}{(integer) NES cutoff to select validated selection-enriched
pathways (default=0.3)}

\item{unenrichNES}{(integer) NES cutoff to select unenriched pathways
(default=0.1)}

\item{enrichDir}{(char) path to directory to store output files
(PLINK files per validated selection-enriched pathway)}

\item{unenrichDir}{(char) path to directory to store output files
(PLINK files per unenriched pathway)}
}
\value{
none
}
\description{
Generates SNP lists per selection-enriched and unenriched pathway
as determined by GSEA
}
